chore: move joblib export to bt_storage function
